Support setting a color only on top or bottom led for mk2
Add the support to set a color only on top or bottom led of a blink(1)
mk2 by adding '@' and the number of the led to the color string. That
can be used in ScriptService by adding '@' and the led number in a text
fiile or by adding a "led" attribute and the led number value to a JSON.

Multi blink(1)s should be supported in ScriptService now, but I cannot
test it.